RAFT: implement client communication, entries transfer, log repository mock, replicated state machine (RSM) abstraction

+++ src/main/java/exchange/core2/revelator/raft/CustomCommandResponse.java
@@ -2,7 +2,7 @@ package exchange.core2.revelator.raft;
 
 import java.nio.ByteBuffer;
 
-public record CustomCommandResponse(long hash, int leaderNodeId, boolean success) implements RpcResponse {
+public record CustomCommandResponse(int hash, int leaderNodeId, boolean success) implements RpcResponse {
 
     @Override
     public int getMessageType() {
@@ -11,14 +11,14 @@ public record CustomCommandResponse(long hash, int leaderNodeId, boolean success
 
     @Override
     public void serialize(ByteBuffer buffer) {
-        buffer.putLong(hash);
+        buffer.putInt(hash);
         buffer.putInt(leaderNodeId);
         buffer.put(success ? (byte) 1 : (byte) 0);
     }
 
     public static CustomCommandResponse create(ByteBuffer buffer) {
 
-        final long hash = buffer.getLong();
+        final int hash = buffer.getInt();
         final int leaderNodeId = buffer.getInt();
         final boolean success = buffer.get() == 1;
